feat(26/sl-pxw5): satsuma-core — NL @-ref extraction and resolution
Add satsuma-core/src/nl-ref.ts with extractBacktickRefs, classifyRef,
resolveRef, extractNLRefData, resolveAllNLRefs, isSchemaInMappingSources.
Uses DefinitionLookup callback (ADR-006) to stay decoupled from WorkspaceIndex.

CLI nl-ref-extract.ts becomes a shim that bridges WorkspaceIndex to core
callbacks. 85 satsuma-core unit tests; all 866 CLI tests pass.
